Add tag fields to delete task pipeline packager.
